split invoking and folding helpers into separate files
