+++ include/alloy/model.hpp
@@ -10,6 +10,49 @@
 #include "external.hpp"
 
 namespace alloy::detail {
+    template<typename F, typename = valid_t>
+    struct invocable {
+        F f;
+
+        constexpr invocable() = delete;
+        constexpr invocable(invocable&&) = default;
+        constexpr invocable(invocable const&) = default;
+
+        constexpr invocable(F&& f)
+            : f{static_cast<F&&>(f)}
+        {}
+
+        template<typename... Args>
+        constexpr decltype(auto) operator ()(Args&&... args) & {
+            return detail::invoke(static_cast<F&>(f), static_cast<Args&&>(args)...);
+        }
+
+        template<typename... Args>
+        constexpr decltype(auto) operator ()(Args&&... args) const& {
+            return detail::invoke(static_cast<F const&>(f), static_cast<Args&&>(args)...);
+        }
+
+        template<typename... Args>
+        constexpr decltype(auto) operator ()(Args&&... args) && {
+            return detail::invoke(static_cast<F&&>(f), static_cast<Args&&>(args)...);
+        }
+
+        template<typename... Args>
+        constexpr decltype(auto) operator ()(Args&&... args) const&& {
+            return detail::invoke(static_cast<F const&&>(f), static_cast<Args&&>(args)...);
+        }
+    };
+
+    template<typename F>
+    struct invocable<F, requires<inheritable<F>>> : F {
+        using base = F;
+        using base::base;
+
+        constexpr invocable(F&& f)
+            : base{static_cast<F&&>(f)}
+        {}
+    };
+
     template<typename S>
     struct source : invocable<S> {
         using base = invocable<S>;
